+++ QRGameTeam15/app/src/main/AndroidManifest.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.qrgameteam15">
-
     <!-- Add permission to use camera and autofocus -->
     <uses-permission android:name="android.permission.CAMERA" />
-    <uses-feature android:name="android.hardware.camera"/>
-    <uses-feature android:name="android.hardware.camera.autofocus"/>
 
+    <uses-feature android:name="android.hardware.camera" />
+    <uses-feature android:name="android.hardware.camera.autofocus" />
 
     <application
         android:allowBackup="true"
@@ -15,6 +14,18 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.QRGameTeam15">
+        <activity
+            android:name=".TakePhoto"
+            android:exported="false" />
+        <activity
+            android:name=".UserMenu"
+            android:exported="false" />
+        <activity
+            android:name=".ExistingUser"
+            android:exported="false" />
+        <activity
+            android:name=".NewUser"
+            android:exported="false" />
         <activity
             android:name=".ScannerView2"
             android:exported="false" />
